Convert distance to a 2 decimal point stright in get_soil.py
We're actually doing a calculation in this script which returns a long float
number. When you use that number, it doesn't match what's in the database which
is a 2 decimal point number. So this limits it to 2 points and also converts it
to a string to ensure it matches what we should find in the production.

Also include notes for matching purposes.

This should finally fix #507 properly!