Add Sentry.io reporting
This adds support for https://sentry.io so that we can manage our application
errors in a more sane fashion. For it to actually work, you need to setup
projects on an account which we'll do on the OSL side. But for general
development this doesn't require anything.

You will need to rebuild your web docker image to test and use this since we
import the raven (Sentry's python library) in settings.py.